+++ BE/src/main/java/com/newsainturtle/shadowmate/follow/controller/FollowController.java
@@ -3,11 +3,11 @@ package com.newsainturtle.shadowmate.follow.controller;
 import com.newsainturtle.shadowmate.auth.service.AuthService;
 import com.newsainturtle.shadowmate.common.BaseResponse;
 import com.newsainturtle.shadowmate.config.auth.PrincipalDetails;
-import com.newsainturtle.shadowmate.follow.dto.AddFollowRequest;
-import com.newsainturtle.shadowmate.follow.dto.DeleteFollowRequestRequest;
-import com.newsainturtle.shadowmate.follow.dto.DeleteFollowerRequest;
-import com.newsainturtle.shadowmate.follow.dto.ReceiveFollowRequest;
-import com.newsainturtle.shadowmate.follow.dto.DeleteFollowingRequest;
+import com.newsainturtle.shadowmate.follow.dto.request.AddFollowRequest;
+import com.newsainturtle.shadowmate.follow.dto.request.DeleteFollowRequestRequest;
+import com.newsainturtle.shadowmate.follow.dto.request.DeleteFollowerRequest;
+import com.newsainturtle.shadowmate.follow.dto.request.ReceiveFollowRequest;
+import com.newsainturtle.shadowmate.follow.dto.request.DeleteFollowingRequest;
 import com.newsainturtle.shadowmate.follow.service.FollowServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
