+++ BE/src/main/java/com/newsainturtle/shadowmate/follow/dto/response/AddFollowResponse.java
@@ -1,4 +1,4 @@
-package com.newsainturtle.shadowmate.follow.dto;
+package com.newsainturtle.shadowmate.follow.dto.response;
 
 import com.newsainturtle.shadowmate.user.enums.PlannerAccessScope;
 import lombok.Builder;
@@ -7,8 +7,6 @@ import lombok.Getter;
 @Getter
 @Builder
 public class AddFollowResponse {
-
     private Long followId;
-
     private PlannerAccessScope plannerAccessScope;
 }
