+++ BE/src/main/java/com/newsainturtle/shadowmate/follow/dto/response/CountFollowResponse.java
@@ -1,4 +1,4 @@
-package com.newsainturtle.shadowmate.follow.dto;
+package com.newsainturtle.shadowmate.follow.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,8 +6,6 @@ import lombok.Getter;
 @Getter
 @Builder
 public class CountFollowResponse {
-
     private Long followerCount;
-
     private Long followingCount;
 }
