+++ BE/src/main/java/com/newsainturtle/shadowmate/follow/dto/response/FollowRequestResponse.java
@@ -1,4 +1,4 @@
-package com.newsainturtle.shadowmate.follow.dto;
+package com.newsainturtle.shadowmate.follow.dto.response;
 
 import com.newsainturtle.shadowmate.user.enums.PlannerAccessScope;
 import lombok.Builder;
@@ -7,18 +7,11 @@ import lombok.Getter;
 @Getter
 @Builder
 public class FollowRequestResponse {
-
     private Long followRequestId;
-
     private Long requesterId;
-
     private String email;
-
     private String nickname;
-
     private String profileImage;
-
     private String statusMessage;
-
     private PlannerAccessScope plannerAccessScope;
 }
