+++ BE/src/main/java/com/newsainturtle/shadowmate/follow/dto/response/FollowerResponse.java
@@ -1,4 +1,4 @@
-package com.newsainturtle.shadowmate.follow.dto;
+package com.newsainturtle.shadowmate.follow.dto.response;
 
 import com.newsainturtle.shadowmate.follow.enums.FollowStatus;
 import lombok.Builder;
@@ -7,18 +7,11 @@ import lombok.Getter;
 @Getter
 @Builder
 public class FollowerResponse {
-
     private Long followId;
-
     private Long followerId;
-
     private String email;
-
     private String nickname;
-
     private String profileImage;
-
     private String statusMessage;
-
     private FollowStatus isFollow;
 }
