+++ BE/src/main/java/com/newsainturtle/shadowmate/follow/service/FollowService.java
@@ -1,6 +1,6 @@
 package com.newsainturtle.shadowmate.follow.service;
 
-import com.newsainturtle.shadowmate.follow.dto.*;
+import com.newsainturtle.shadowmate.follow.dto.response.*;
 import com.newsainturtle.shadowmate.follow.enums.FollowStatus;
 import com.newsainturtle.shadowmate.user.entity.User;
 
@@ -9,22 +9,13 @@ import java.util.List;
 public interface FollowService {
 
     List<FollowingResponse> getFollowing(final User user);
-
     List<FollowerResponse> getFollower(final User user);
-
     List<FollowRequestResponse> getFollowRequestList(final User user);
-
     AddFollowResponse addFollow(final User user, final Long targetUserId);
-
     void deleteFollowing(final User user, final Long targetUserId);
-
     void deleteFollower(final User user, final Long targetUserId);
-
     void deleteFollowRequest(final User user, final Long targetUserId);
-
     String receiveFollow(final User user, final Long targetUserId, final boolean followReceive);
-
     FollowStatus isFollow(final User user, final User searchUser);
-
     CountFollowResponse countFollow(final Long userId);
 }
